Move and update query tx utils to different package so CLI can use them
